Add support for naming metrics (#21)
This commit adds support for optionally naming metrics. Previously, the
instance name was used for the metric name when publishing metrics to
Wavefront; however, this can create problems while populating a
integration dashboard as the instance name holds the namespace to which
the instance belongs. If an instance was defined in a non-istio-system
namespace, it wouldn't have been captured by the integration.

Signed-off-by: Venil Noronha <veniln@vmware.com>